+++ src/com/geoscope/GeoEye/Space/TypesSystem/CoComponent/ObjectModel/GeoMonitoredObject1/DEVICE/VideoRecorderModule/TVideoRecorderServerViewUDPRTP.java
@@ -572,7 +572,7 @@ public class TVideoRecorderServerViewUDPRTP extends TVideoRecorderServerView {
 	private TLANConnectionUDPRepeater	VideoLocalServer = null;
 	private TVideoClient				VideoClient = null;
 	
-    public TVideoRecorderServerViewUDPRTP(Context pcontext, String pGeographProxyServerAddress, int pGeographProxyServerPort, int pUserID, String pUserPassword, TCoGeoMonitorObject pObject, boolean pflAudio, boolean pflVideo, String pUserAccessKey, TExceptionHandler pExceptionHandler, TextView plbVideoRecorderServer) {
+    public TVideoRecorderServerViewUDPRTP(Context pcontext, String pGeographProxyServerAddress, int pGeographProxyServerPort, long pUserID, String pUserPassword, TCoGeoMonitorObject pObject, boolean pflAudio, boolean pflVideo, String pUserAccessKey, TExceptionHandler pExceptionHandler, TextView plbVideoRecorderServer) {
     	super(pcontext, pGeographProxyServerAddress,pGeographProxyServerPort, pUserID,pUserPassword, pObject, pflAudio,pflVideo, pUserAccessKey, pExceptionHandler, plbVideoRecorderServer);
 		if (GeographProxyServerPort == TUDPEchoServerClient.ServerDefaultPort) {
 	    	TTracker Tracker = TTracker.GetTracker();
