+++ src/com/geoscope/GeoEye/Space/TypesSystem/DataStream/TDataStreamFunctionality.java
@@ -60,7 +60,7 @@ public class TDataStreamFunctionality extends TComponentFunctionality {
 	public byte[] GetDescriptorData() throws Exception {
 		String URL1 = Server.Address;
 		//. add command path
-		URL1 = "http://"+URL1+"/"+"Space"+"/"+"2"/*URLProtocolVersion*/+"/"+Integer.toString(Server.User.UserID);
+		URL1 = "http://"+URL1+"/"+"Space"+"/"+"2"/*URLProtocolVersion*/+"/"+Long.toString(Server.User.UserID);
 		String URL2 = "TypesSystem"+"/"+Integer.toString(TypeFunctionality.idType)+"/"+"Co"+"/"+Long.toString(idComponent)+"/"+"Data.dat";
 		//. add command parameters
 		URL2 = URL2+"?"+"0"/*command*/+","+"1"/*Parameters version*/;
